car_rental/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
